+++ R/data-woodhouse.R
@@ -1,0 +1,17 @@
+#' Paleo reconstructions of Lees Ferry flow
+#' 
+#' There are various paleo reconstructions of Lees Ferry flow. Two of these 
+#' data sets are included in this package for ease of use in CRSS: Woodhouse 
+#' et al. (2006) and Meko et al. (2017). 
+#' 
+#' @format `woodhouse` contains an [nfd] object containing Woodhouse et al. 
+#'   2006's "Lees-B" reconstruction. This includes Lees Ferry flow reconstructed 
+#'   for water years 1490-1998. 
+#'   
+#' @source [Woodhouse, C.A., S.T. Gray, and D.M. Meko, 2006. Updated streamflow 
+#' reconstructions for the Upper Colorado River basin.  Water Resources 
+#' Research.](https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2005WR004455)
+#' 
+#' @rdname paleo_data
+#' 
+"woodhouse"