+++ NEWS.md
@@ -9,6 +9,7 @@ CRSSIO 0.8.2.9000
   - added internal data from Meko et al. 2018's paleo reconstruction of the 4-river volume for use in this function. 
   - also updated internal historical index data to go through 2019. 
 - added `as_nfd.data.frame()`, `as_nfd.list()`, `as_crss_nf.data.frame()`, and `as_crss_nf.list()` to convert data frames to `nfd` and `crss_nf` objects. (#82)
+- added Meko et al. (2017) and Woodhouse et al. (2006) Lees Ferry paleo reconstrutions for use in package. (#114)
 
 CRSSIO 0.8.2
 =================